% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DepInterPackages.R
\name{add.html.help}
\alias{add.html.help}
\title{Given html help for a function}
\usage{
add.html.help(package, func, tempsave = paste0(getwd(), "/temp.html"))
}
\arguments{
\item{package}{: Names of package to extract help}

\item{func}{: Name of function to extract help}

\item{tempsave}{: Tempory file to save help}
}
\description{
Given html help for a function
}

