% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkAll.R
\name{allDepFunction}
\alias{allDepFunction}
\title{For a function, give all dependencies}
\usage{
allDepFunction(envir, name.functions)
}
\arguments{
\item{envir}{: environments where the function should search dependencies.}

\item{name.function}{: function name (character)}
}
\value{
object of class dependenciesGraphs
}
\description{
For a function, give all dependencies
}
