% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{AllFunctionEnv}
\alias{AllFunctionEnv}
\title{Given a environment, can be a package, returns a vector of fonctions includes in this environment}
\usage{
AllFunctionEnv(Envi)
}
\arguments{
\item{Envi}{: Environment, try fonction "search()" to obtain a list of charged environments}
}
\value{
nomfun : Names of fonctions includes in this environment.
}
\description{
Given a environment, can be a package, returns a vector of fonctions includes in this environment
}

