% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sourcescript.R
\name{data.graph.script}
\alias{data.graph.script}
\title{Dependencies for scripts}
\usage{
data.graph.script(path)
}
\arguments{
\item{path}{: path of R Script}
}
\description{
Dependencies for scripts
}
\examples{
dep <- data.graph.script('MyRScrpt.R')
plot(dep)
}

