% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{funDependencies}
\alias{funDependencies}
\title{Return all dependencies from a function in an environment}
\usage{
funDependencies(envir, name.function)
}
\arguments{
\item{envir}{: environment where you want to scherch dependencies}

\item{name.function}{: Function name (in character)}
}
\value{
List with nodes and edges informations. Needed for visNetwork visualization.
}
\description{
Return all dependencies from a function in an environment
}
\examples{
dep <- funDependencies("package:ibr","iterchoiceS1")
plot(dep)
}

