% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{launch.app}
\alias{launch.app}
\title{Launch shiny app}
\usage{
launch.app(path.install.package = NULL)
}
\arguments{
\item{path.install.package}{: Where you have install package DependenciesGraphs. If null, search in current library}
}
\description{
Launch shiny app
}
