% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{linksForAll}
\alias{linksForAll}
\title{For a environnement, give all dependencies}
\usage{
linksForAll(envir)
}
\arguments{
\item{envir}{: environment where the function should search dependencies.}
}
\value{
Dataframe with two columns, 'master'and 'slave'
}
\description{
For a environnement, give all dependencies
}

