% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{LinksForOne}
\alias{LinksForOne}
\title{For one fonction, give all dependencies}
\usage{
\code{require("mvbutils")}
\code{LinksForOne(Enviro="package:mvbutils",Nomfonciton="foodweb")}
}
\arguments{
\item{Enviro}{: environment where the function should search dependencies.}

\item{Nomfonciton}{: function name (character)}
}
\value{
ancdsc : Dataframe which two colums, first : master, second : slave
}
\description{
For one fonction, give all dependencies
}

