% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{mastersSlaves}
\alias{mastersSlaves}
\title{Given a matrix, returns a dataframe which two columns 'master'and 'slave'}
\usage{
mastersSlaves(Mat)
}
\arguments{
\item{Mat}{: matrix of dependencies, should be named (cols and rows), should be square matrix, dependencies are noted 1.}
}
\value{
Dataframe with two columns 'master'and 'slave'
}
\description{
Given a matrix, returns a dataframe which two columns 'master'and 'slave'
}

