% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepInterPackages.R
\name{Pck.load}
\alias{Pck.load}
\title{Given dependencies between all install packages}
\usage{
Pck.load()
}
\value{
List of dependencies
}
\description{
Given dependencies between all install packages
}
