% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{plot.dependenciesGraph}
\alias{plot.dependenciesGraph}
\title{Plot network for dependenciesGraph object}
\usage{
\method{plot}{dependenciesGraph}(object)
}
\arguments{
\item{object}{: dependenciesGraph object.}
}
\description{
Plot network for dependenciesGraph object. Using visNetwork package.
}
\details{
Plot network for dependenciesGraph object. Using visNetwork package.
}
\examples{
dep <- funDependencies("package:ibr","iterchoiceS1")
plot(dep)
}

