% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{plot.dependenciesGraphs}
\alias{plot.dependenciesGraphs}
\title{Plot network for dependenciesGraphs object}
\usage{
\method{plot}{dependenciesGraphs}(object, block = FALSE, width = "100\%",
  height = "400px")
}
\arguments{
\item{object}{: dependenciesGraphs object.}

\item{width}{: Width (optional, defaults to automatic sizing)}

\item{height}{: Height (optional, defaults to automatic sizing)}
}
\description{
Plot network for dependenciesGraphs object. Using visNetwork package.
}
\details{
Plot network for dependenciesGraphs object. Using visNetwork package.
}
\examples{

dep <- funDependencies('package:ibr','iterchoiceS1')
plot(dep)

# size
plot(dep, height = "800px", width = "100\%")

}
