% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{PrepareToVis}
\alias{PrepareToVis}
\title{Prepare data for graph visNetwork}
\usage{
PrepareToVis(liens, Nomfonctions = NULL)
}
\arguments{
\item{Nomfonctions}{: All fonctions to includes in graph (default : union(masters & slaves))}

\item{Liens}{: Dataframe, two colums, the first is master, the second is slaves}
}
\value{
visdata : List contain elements tu do a visNetwork graph
}
\description{
Prepare data for graph visNetwork
}

