% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{VisFun}
\alias{VisFun}
\title{Return all dependencies between a function and others functions in an envorinnement}
\usage{
VisFun(Enviro, Func)
}
\arguments{
\item{Enviro}{: envorinnement where you want to scherch dependencies}

\item{Func}{: Function name (in character)}
}
\value{
visdata : List contain elements tu do a visNetwork graph
}
\description{
Return all dependencies between a function and others functions in an envorinnement
}

