% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{VisFuns}
\alias{VisFuns}
\title{Return all dependencies between fonctions in an envorinnement}
\usage{
VisFuns(Enviro)
}
\arguments{
\item{Enviro}{: envorinnement where you want to scherch dependencies}
}
\value{
visdata : List contain elements to do a visNetwork graph
}
\description{
Return all dependencies between fonctions in an envorinnement
}

