% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{VisFunsmatrice}
\alias{VisFunsmatrice}
\title{Return all dependencies between elements of a matrix}
\usage{
VisFunsmatrice(Mat)
}
\arguments{
\item{Mat}{: matrix of dependencies, should be named (cols and rows), should be square matrix, dependencies are noted 1.}
}
\value{
visdata : List contain elements to do a visNetwork graph
}
\description{
Return all dependencies between elements of a matrix
}

