% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{allFunctionEnv}
\alias{allFunctionEnv}
\title{Given a environment, can be a package, returns a vector of included functions in this environment}
\usage{
allFunctionEnv(name, envir)
}
\arguments{
\item{envir}{: Environment, try "search()" to obtain a list of loaded environments}
}
\value{
Names of included functions in this environment.
}
\description{
Given a environment, can be a package, returns a vector of included functions in this environment
}

