% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{envirDependencies}
\alias{envirDependencies}
\title{Return all dependencies between functions in an environment}
\usage{
envirDependencies(envir)
}
\arguments{
\item{envir}{: environment on which you want to search dependencies}
}
\value{
List with nodes and edges informations. Needed for visNetwork visualization.
}
\description{
Return all dependencies between functions in an environment
}
\examples{
dep <- envirDependencies("package:ibr")
plot(dep)
}

