% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{linksForOne}
\alias{linksForOne}
\title{For a function, give all dependencies}
\usage{
linksForOne(envir, name.function)
}
\arguments{
\item{envir}{: environment where the function should search dependencies.}

\item{name.function}{: function name (character)}
}
\value{
Dataframe with two columns, 'master'and 'slave'
}
\description{
For a function, give all dependencies
}

