% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{MastersSlaves}
\alias{MastersSlaves}
\title{Given a matrix, returns a dataframe which two colums, first : master, second : slave}
\usage{
\code{require("mvbutils")}
\code{graphfun<-foodweb( where="package:mvbutils",prune="foodweb",descendents=T,plotting =F,ancestors=F)$funmat}
\code{MastersSlaves(graphfun)}
}
\arguments{
\item{Mat}{: matrix of dependencies, should be named (cols and rows), should be square matrix, dependencies are noted 1.}
}
\value{
M_s : Dataframe which two colums, first : master, second : slave
}
\description{
Given a matrix, returns a dataframe which two colums, first : master, second : slave
}

