% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DepInterPackages.R
\name{Pck.load.to.vis}
\alias{Pck.load.to.vis}
\title{Given dependencies between all install packages}
\usage{
Pck.load.to.vis(Packages = "All", color.table = c("red", "blue", "green",
  "yellow", "pink"))
}
\arguments{
\item{Packages}{: Names of packages to includes to extract}

\item{color.table}{: Name of color to use}
}
\value{
List to graph
}
\description{
Given dependencies between all install packages
}

