% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepInterPackages.R
\name{Pck.load.to.vis}
\alias{Pck.load.to.vis}
\title{Given dependencies between all install packages}
\usage{
Pck.load.to.vis(Packages = "All", color.table = c("#0B0B3B", "#0404B4",
  "#5858FA", "#A9A9F5"))
}
\arguments{
\item{Packages}{: Names of packages to includes to extract}

\item{color.table}{: Name of color to use}
}
\value{
List to graph
}
\description{
Given dependencies between all install packages
}
