% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{plot.dependenciesGraphs}
\alias{plot.dependenciesGraphs}
\title{Plot network for dependenciesGraphs object}
\usage{
\method{plot}{dependenciesGraphs}(object, block = FALSE)
}
\arguments{
\item{object}{: dependenciesGraphs object.}
}
\description{
Plot network for dependenciesGraphs object. Using visNetwork package.
}
\details{
Plot network for dependenciesGraphs object. Using visNetwork package.
}
\examples{
dep <- funDependencies("package:ibr","iterchoiceS1")
plot(dep)
}

