% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/VisGraphAuto.R
\name{prepareToVis}
\alias{prepareToVis}
\title{Prepare data for graph visNetwork}
\usage{
prepareToVis(link, functions = NULL)
}
\arguments{
\item{link}{: Dataframe, two colums, the first is master, the second is slaves}

\item{functions}{: All functions to includes in graph (default : union(masters & slaves))}
}
\value{
List contain elements needed for visNetwork visualization
}
\description{
Prepare data for graph visNetwork
}

